# backend: improve parse function performance in gtime package