--- a/backend/gtime/gtime.go
+++ b/backend/gtime/gtime.go
@@ -75,18 +75,45 @@ func ParseDuration(inp string) (time.Duration, error) {
 }
 
 func parse(inp string) (time.Duration, string, error) {
-	result := dateUnitPattern.FindSubmatch([]byte(inp))
+	// Fast path for simple duration formats (no date units)
+	if len(inp) > 0 {
+		lastChar := inp[len(inp)-1]
+		if lastChar != 'd' && lastChar != 'w' && lastChar != 'M' && lastChar != 'y' {
+			dur, err := time.ParseDuration(inp)
+			return dur, "", err
+		}
+
+		// Check if the rest is a number for date units
+		numPart := inp[:len(inp)-1]
+		isNum := true
+		for _, c := range numPart {
+			if c < '0' || c > '9' {
+				isNum = false
+				break
+			}
+		}
+		if isNum {
+			num, err := strconv.Atoi(numPart)
+			if err != nil {
+				return 0, "", err
+			}
+			return time.Duration(num), string(lastChar), nil
+		}
+	}
+
+	// Fallback to regex for complex cases
+	result := dateUnitPattern.FindStringSubmatch(inp)
 	if len(result) != 3 {
 		dur, err := time.ParseDuration(inp)
 		return dur, "", err
 	}
 
-	num, err := strconv.Atoi(string(result[1]))
+	num, err := strconv.Atoi(result[1])
 	if err != nil {
 		return 0, "", err
 	}
 
-	return time.Duration(num), string(result[2]), nil
+	return time.Duration(num), result[2], nil
 }
 
 // FormatInterval converts a duration into the units that Grafana uses