**What this PR does / why we need it**:

`parse` function is used in `ParseDuration` and `ParseInterval`. Those methods is used in many places in `grafana/grafana`. `ParseDuration` is especially used in settings.go to parse the config. Improving their performance will result a better start time and less memory usage. 

This is also a complimentary PR for https://github.com/grafana/grafana-plugin-sdk-go/pull/1237

Here is the benchmark comparison:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/grafana-plugin-sdk-go/backend/gtime
cpu: Apple M1 Pro
                     │  pmem.0.txt   │             pmem.2.txt             │
                     │    sec/op     │   sec/op     vs base               │
Parse/PureNumber-10    164.50n ± 11%   91.89n ± 1%  -44.14% (p=0.002 n=6)
Parse/SimpleUnit-10     79.85n ±  0%   15.21n ± 0%  -80.94% (p=0.002 n=6)
Parse/ComplexUnit-10   105.80n ±  0%   42.40n ± 1%  -59.93% (p=0.002 n=6)
Parse/DateUnit-10      148.55n ±  1%   13.27n ± 4%  -91.07% (p=0.002 n=6)
Parse/MonthUnit-10     151.40n ±  1%   13.66n ± 1%  -90.97% (p=0.002 n=6)
Parse/YearUnit-10      151.20n ±  1%   13.59n ± 0%  -91.01% (p=0.002 n=6)
geomean                 129.5n         22.95n       -82.28%

                     │  pmem.0.txt  │               pmem.2.txt               │
                     │     B/op     │    B/op     vs base                    │
Parse/PureNumber-10      80.00 ± 0%   72.00 ± 0%   -10.00% (p=0.002 n=6)
Parse/SimpleUnit-10      8.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
Parse/ComplexUnit-10     8.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
Parse/DateUnit-10      152.000 ± 0%   4.000 ± 0%   -97.37% (p=0.002 n=6)
Parse/MonthUnit-10     152.000 ± 0%   4.000 ± 0%   -97.37% (p=0.002 n=6)
Parse/YearUnit-10      152.000 ± 0%   4.000 ± 0%   -97.37% (p=0.002 n=6)
geomean                  51.18                    ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                     │ pmem.0.txt │               pmem.2.txt               │
                     │ allocs/op  │ allocs/op   vs base                    │
Parse/PureNumber-10    5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.002 n=6)
Parse/SimpleUnit-10    1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
Parse/ComplexUnit-10   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
Parse/DateUnit-10      3.000 ± 0%   1.000 ± 0%   -66.67% (p=0.002 n=6)
Parse/MonthUnit-10     3.000 ± 0%   1.000 ± 0%   -66.67% (p=0.002 n=6)
Parse/YearUnit-10      3.000 ± 0%   1.000 ± 0%   -66.67% (p=0.002 n=6)
geomean                2.265                    ?                      ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```